// src/models/mst_user_module.ts
class MstUserModule {
    id?:            number = 0;
    user_id:        number = 0;
    Module_id:      number = 0;
    can_open:       boolean = false;
    can_add:        boolean = false;
    can_update:     boolean = false;
    can_lock:       boolean = false;
    can_unlock:     boolean = false;
    can_delete:     boolean = false;
    can_print:      boolean = false;

}

export default MstUserModule;